Partly fixed datapack integration
