+++ include/datagui/gui.hpp
@@ -14,6 +14,7 @@
 #include "datagui/element/series.hpp"
 #include "datagui/element/text_box.hpp"
 #include "datagui/element/text_input.hpp"
+#include "datapack/datapack.hpp"
 
 namespace datagui {
 
@@ -100,6 +101,12 @@ public:
     return tree.variable<T>([initial_value]() { return initial_value; });
   }
 
+  // Datapack
+
+  template <typename T>
+  requires datapack::writeable<T> && datapack::readable<T>
+  Variable<T> edit_variable(const T& initial_value = T());
+
   // Style
 
   void style(const Style& style) {
@@ -138,5 +145,95 @@ private:
   Element element_hover;
   int next_float_priority = 0;
 };
+
+class GuiWriter : public datapack::Writer {
+public:
+  GuiWriter(Gui& gui) : gui(gui) {}
+
+  void number(datapack::NumberType type, const void* value) override;
+  void boolean(bool value) override;
+  void string(const char*) override;
+  void enumerate(int value, const std::span<const char*>& labels) override;
+  void binary(const std::span<const std::uint8_t>& data) override;
+
+  void optional_begin(bool has_value) override;
+  void optional_end() override;
+
+  void variant_begin(int value, const std::span<const char*>& labels) override;
+  void variant_end() override;
+
+  void object_begin() override;
+  void object_next(const char* key) override;
+  void object_end() override;
+
+  void tuple_begin() override;
+  void tuple_next() override;
+  void tuple_end() override;
+
+  void list_begin() override;
+  void list_next() override;
+  void list_end() override;
+
+private:
+  Gui& gui;
+  std::stack<Variable<std::size_t>> list_sizes;
+};
+
+class GuiReader : public datapack::Reader {
+public:
+  GuiReader(Gui& gui) : gui(gui) {}
+
+  void number(datapack::NumberType type, void* value) override;
+  bool boolean() override;
+  const char* string() override;
+  int enumerate(const std::span<const char*>& labels) override;
+  std::span<const std::uint8_t> binary() override;
+
+  bool optional_begin() override;
+  void optional_end() override;
+
+  int variant_begin(const std::span<const char*>& labels) override;
+  void variant_end() override;
+
+  void object_begin() override;
+  void object_next(const char* key) override;
+  void object_end() override;
+
+  void tuple_begin() override;
+  void tuple_next() override;
+  void tuple_end() override;
+
+  void list_begin() override;
+  bool list_next() override;
+  void list_end() override;
+
+private:
+  Gui& gui;
+  std::stack<std::pair<Variable<std::size_t>, std::size_t>> list_sizes;
+};
+
+template <typename T>
+requires datapack::writeable<T> && datapack::readable<T>
+Variable<T> Gui::edit_variable(const T& initial_value) {
+  series_begin_force();
+  auto var = variable<T>(initial_value);
+  if (var.modified() || var.is_new()) {
+    printf("Var modified? %s\n", var.modified() ? "true" : "false");
+    printf("Var new? %s\n", var.is_new() ? "true" : "false");
+    printf("Writing new value\n");
+    series_begin_force();
+    GuiWriter(*this).value(*var);
+    series_end();
+  } else {
+    if (series_begin()) {
+      printf("Revisit, re-trigger variable\n");
+      GuiReader(*this).value(var.mut());
+      var.trigger();
+      series_end();
+    }
+  }
+  series_end();
+  return var;
+}
 
 } // namespace datagui
